# Remove extra `malloc` from `kRing`