--- a/src/apps/benchmarks/benchmarkKRing.c
+++ b/src/apps/benchmarks/benchmarkKRing.c
@@ -23,14 +23,12 @@ H3Index pentagon = 0x89080000003ffff;
 
 BEGIN_BENCHMARKS();
 
-H3Index* out = malloc(H3_EXPORT(maxKringSize)(400) * sizeof(H3Index));
+H3Index* out = malloc(H3_EXPORT(maxKringSize)(40) * sizeof(H3Index));
 
 BENCHMARK(kRing10, 10000, { H3_EXPORT(kRing)(hex, 10, out); });
 BENCHMARK(kRing20, 10000, { H3_EXPORT(kRing)(hex, 20, out); });
 BENCHMARK(kRing30, 10000, { H3_EXPORT(kRing)(hex, 30, out); });
 BENCHMARK(kRing40, 10000, { H3_EXPORT(kRing)(hex, 40, out); });
-BENCHMARK(kRing100, 1000, { H3_EXPORT(kRing)(hex, 100, out); });
-BENCHMARK(kRing400, 1000, { H3_EXPORT(kRing)(hex, 400, out); });
 
 BENCHMARK(kRingPentagon10, 500, { H3_EXPORT(kRing)(pentagon, 10, out); });
 BENCHMARK(kRingPentagon20, 500, { H3_EXPORT(kRing)(pentagon, 20, out); });