The fast path of `kRing`, doesn't need a `malloc`, and these changes remove it.

See #319.