@isaacbrodsky 2020-03-22
> I couldn't figure out how to run benchmarks for `kRing`. Does anyone know how?

When running CMake, you need to set `-DBUILD_BENCHMARKS=ON`.  Then after running your make process, you can run `./bin/benchmarkKRing`