@isaacbrodsky 2020-03-22
Looks like the failing CI checks are formatting (needs `make format` to be run) and Mac OSX (looks like a transient failure)