@ajfriend 2020-03-22
Just for my future notes:

```
make clean
cmake . -DBUILD_BENCHMARKS=ON
make
make test
./bin/benchmarkKRing
```