@ajfriend 2020-03-22
Here are some timing benchmarks comparing the old algorithm to the new one.

|                 |        old (μs) |        new (μs) |   % change |
|:----------------|-----------:|-----------:|-----------:|
| kRing10         |    13.7782 |    13.645  |      -0.97 |
| kRing20         |    50.593  |    50.5054 |      -0.17 |
| kRing30         |   112.482  |   110.635  |      -1.64 |
| kRing40         |   195.406  |   193.416  |      -1.02 |
| kRing100        |  1178.59   |  1171.8    |      -0.58 |
| kRing400        | 17910.4    | 17707.1    |      -1.14 |
| kRingPentagon10 |   393.074  |   384.726  |      -2.12 |
| kRingPentagon20 |  3126.3    |  3135.22   |       0.29 |
| kRingPentagon30 | 10493.8    | 10355.7    |      -1.32 |
| kRingPentagon40 | 24282.6    | 24226.8    |      -0.23 |