@isaacbrodsky 2020-03-22
> 
> 
> > Looks like the failing CI checks are formatting (needs `make format` to be run) and Mac OSX (looks like a transient failure)
> 
> ```
> $ make format
> make: *** No rule to make target `format'.  Stop.
> ```
> 
> And I didn't see any other likely targets in the `Makefile`. Maybe a different command?

I believe this happens if you don't have `clang-format` installed when you run CMake. Does CMake issue a warning about not finding `clang-format`? (I think on Mac OSX, it can be installed with `brew install clang-format`)